% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2014}
\alias{accident_2014}
\title{Fatal injuries suffered in motor vehicle traffic crashes in 2014}
\format{A data frame with columns:}
\usage{
accident_2014
}
\description{
It is a data from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System, which is a nationwide census providing
the American public yearly data regarding fatal injuries suffered in motor
vehicle traffic crashes.
}
\examples{
\dontrun{
 accident_2014
}
}
\keyword{datasets}
