% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the accidents in states in the fars data.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number which represent a state.}

\item{year}{a number which represent a year.}
}
\value{
It returns a plot of state with points where the accidents
occured in a specific year.
}
\description{
This function plots the state and the point where the accidents
occured in a specific year.
}
\note{
It plots the accidents which occured in latitude <= 90 and
longitude <= 900. If the argument \code{state.num} is beyond the number of
states in the fars data then it will return that "invalid STATE number".
If there is no accident during a specific year in  aspecific state then
it returns "no accidents to plot".
}
\examples{

\dontrun{
fars_map_state(1,2013)
}

}
