% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of a .csv file.}
}
\value{
It returns a  \code{tbl} data frame object.
}
\description{
This function reads a .csv file and convert in to \code{tbl} data
frame object. The main advantage to using a tbl_df over a regular data frame is
the printing: tbl objects only print a few rows and all the columns that fit on
one screen, describing the rest of it as text.
}
\note{
If there is no such file then it will return "file <filename> does not
exist".
}
\examples{
\dontrun{
fars_read("/home/kanna/Desktop/COURSERA/data/accident_2013.csv.bz2")
}

}
