% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month and year columns in the fars data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of numbers.}
}
\value{
It returns a list of data frame for each year. The data frame contains
only the month and the year columns of fars data.
}
\description{
This function returns a list of data frames. Each data frame contain only
the month and the year columns from fars data for each year.
}
\note{
If the argument is not a number then it will return a NULL list with
a warning says that the argument is invalid.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013,2014))

}
