% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Frequency table for months in fars data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of numbers.}
}
\value{
It returns a frequency table for months in each year. The first column
of the output is the month number and the other columns are year-wise month
counts in the fars data.
}
\description{
This function gives a frequency table for months in each year.
}
\note{
If the year is not in the fars data then it will return an error message.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))

}
