% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate fars data file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a numeric value.}
}
\value{
It returns character strings eg: "accident_2013.csv.bz2".
}
\description{
It generates character strings (eg: "accident_2013.csv.bz2") for
a given year (eg: 2013).
}
\note{
If the argument is not a number then it will return "accident_NA.csv.bz2"
and a warning.
}
\examples{
make_filename(2013)

}
